fix(code/driver): Fix duplicated work by not storing the proposal prior to passing it as input to the driver
We are essentially duplicating work here and making the code harder to
wrap one's head around since we actually do not need to reach directly
into the driver's proposal keeper here.

Indeed, a few lines below we send a `Proposal` input to the driver, which
then calls `mux_proposal` whose first action is to store the proposal
in the proposal keeper.