chore(code): Namespace crates under `informalsystems-` prefix, add `bft` suffix
Rename all crates to include a `informalsystems-` prefix and a `bft` suffix to avoid clashes with the existing `malachite` crate.

So `malachite-consensus` would become `informalsystems-malachitebft-consensus`.

It is a mouthful but that can be alleviated in code by specifying a package alias in `Cargo.toml`:

```toml
[dependencies]
malachitebft-consensus = { version = "0.1.0", package = "informalsystems-malachitebft-consensus" }
```

which allows import of the form:

```rust
use malachitebft_consensus::foobar::*
```

instead of

```rust
use informalsystems_malachitebft_consensus::foobar::*
```