chore(CCHAIN309): Check all scenarios when the driver panics (#1178)
Closes: #XXX

Triggered by the panic observed
[here](https://github.com/circlefin/malachite/pull/1177), we decided to
take a closer look to see when the driver can panic and whether it can
happen as part of regular execution or only due to bugs in our code.

There are 3 scenarios when driver can panic:
1. in `./src/proposal_keeper.rs` when it tries to store two proposals
from two different proposers. This happened in the case linked above,
and we need to make sure that a proposal received from a non-proposer is
dropped before it reaches the driver so this panic can never happen.
```rust
        // Check if the proposal is from the same validator as the first one recorded.
        // This should never happen, as proposals should always come from the same validator (round's proposer).
        if let Some(first) = self.get_first_proposal() {
            if first.validator_address() != proposal.validator_address() {
                panic!(
                    "BUG: Received proposals from different validators in the same round.\n\
                    Existing: {:?}, New: {:?}",
                    first.validator_address(),
                    proposal.validator_address()
                );
            }
        }
```

2. in `./src/driver.rs` we have two panics when dealing with round
certificates. Namely, when the vote keeper generates `Skip` or
`PrecommitAny` certificate event based on the votes it received, the
driver tries to store these certificates in its `round_certificate`
field. Before storing driver checks if votes for these certificates
exist in the vote keeper (actually it checks if vote keeper contains
votes from the same round as certificate), and if not it panics. This
should never happen because the same vote keeper generated these
certificate events in the first place because it has these votes.

```rust

fn store_precommit_any_round_certificate(&mut self, vote_round: Round) {
        let Some(per_round) = self.vote_keeper.per_round(vote_round) else {
            panic!("Missing the PrecommitAny votes for round {vote_round}");
        };

fn store_skip_round_certificate(&mut self, vote_round: Round) {
        let Some(per_round) = self.vote_keeper.per_round(vote_round) else {
            panic!("Missing the SkipRoundvotes for round {vote_round}");
        };

```

In the end, this PR:
- add extra check that `PrecommitAny` is from the current round before
storing it, this is needed to prevent that the current round certificate
is overwritten by an old one
- add a more complete logic for proposal verification, namely the new
code at the app checks if proposer is indeed from the correct proposer
and if it is properly signed.
- In case proposal parts are for a future height it stores them in
`pending` and only process them and verify when we eventually start this
specific height and round parts are for

---

### PR author checklist

#### For all contributors

- [ ] Reference a GitHub issue
- [ ] Ensure the PR title follows the [conventional
commits][conv-commits] spec
- [ ] Add a release note in [`RELEASE_NOTES.md`](/RELEASE_NOTES.md) if
the change warrants it
- [ ] Add an entry in [`BREAKING_CHANGES.md`](/BREAKING_CHANGES.md) if
the change warrants it

#### For external contributors

- [ ] Maintainers of Malachite are [allowed to push changes to the
branch][gh-edit-branch]

[conv-commits]: https://www.conventionalcommits.org/en/v1.0.0/#summary
[gh-edit-branch]:
https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork#enabling-repository-maintainer-permissions-on-existing-pull-requests

---------

Co-authored-by: Romain Ruetschi <github@romac.me>