fix(driver): `PrecommitAny` does not shadow `PolkaNil` and `PolkaAny` pending inputs
Test unit confirming this scenario:

> According to the spec, PolkaNil a Precommit for nil should be
broadcasted whenever a quorum of nil prevotes are observed. However,
`PrecommitAny` might shadow `PolkaNil`
[here](https://github.com/circlefin/malachite/blob/b205f4252f3064d9a74716056f63834ff33f2de9/code/crates/core-driver/src/mux.rs#L397)
(L44 / L47). This prevents the validator from sending out a nil
precommit vote when it should. We currently believe the impact is
minimal since the existence of `PrecommitAny` should be observable by
all validators under the gossip communication model, even if the current
validator skips `PolkaNil` broadcast

The solution is not optimal, since it allows the production of multiple
`PrecommitAny` inputs, therefore the `timeoutPrecommit` is scheduled
multiple times per round, which is annoying but not per se an issue. The
solution consist on having `find_non_value_threshold` returning a vector
of thresholds, instead of a single one. We allow `PolkaNil` to shadow
`PolkaAny`, but none of them is shadowed by `PrecommitAny`.

---

### PR author checklist

#### For all contributors

- [ ] Reference a GitHub issue
- [ ] Ensure the PR title follows the [conventional
commits][conv-commits] spec
- [ ] Add a release note in [`RELEASE_NOTES.md`](/RELEASE_NOTES.md) if
the change warrants it
- [ ] Add an entry in [`BREAKING_CHANGES.md`](/BREAKING_CHANGES.md) if
the change warrants it

#### For external contributors

- [ ] Maintainers of Malachite are [allowed to push changes to the
branch][gh-edit-branch]

[conv-commits]: https://www.conventionalcommits.org/en/v1.0.0/#summary
[gh-edit-branch]:
https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork#enabling-repository-maintainer-permissions-on-existing-pull-requests

---------

Co-authored-by: Romain Ruetschi <romain.ruetschi@circle.com>